Continuing with deployment (mysql incompatibility - get a local instance set up)

+++ database/migrations/2021_10_19_232044_add_author_to_blog_posts_table.php
@@ -14,7 +14,7 @@ class AddAuthorToBlogPostsTable extends Migration
     public function up()
     {
         Schema::table('blog_posts', function (Blueprint $table) {
-            $table->unsignedInteger('user_id')->default(0);
+            $table->unsignedBigInteger('user_id')->default(0);
 
             $table->foreign('user_id')->references('id')->on('users');
         });
